Update interfaces, prefer authorized-keys over public key.
This commit includes two changes. First, if a user building an image
has an authorized_keys file, we will push that to the image. This
means that you will ONLY be able to ssh into the pi running this image
if your authorized_keys file includes the machine which is being used
to build the image.

If an authorized_keys file is not found, the public key on the local
machine is used instead.

A short comment has been added to the interfaces file showing how one
can regenerate the file if a different SSID must be used.